Interact with a trading backend

Interact with a trading backend

trader(
  orders,
  trader_life = 300,
  resubmit_interval = 5,
  trading_connection,
  pricing_connection = NULL,
  pricing_spread_tolerance = 0.01,
  pricing_overrides = NULL,
  verbose = TRUE
)

Arguments

orders

a data frame of orders

trader_life

how long, in seconds, the trader should operate before timeout

resubmit_interval

how long, in seconds, the trader should wait before cancelling all open orders, recompute price limits, and resubmit orders

trading_connection

the backend trading connection for submitting orders

pricing_connection

the backend trading connection for computing price limits

pricing_spread_tolerance

the maximum bid-ask spread which can be tolerated for a limit price to be computed

pricing_overrides

optional data frame of limit prices to assign to each symbol

verbose

TRUE/FALSE should the function be chatty?

Value

a data frame detailing the outcome of each attempted order

Examples

t_conn <- alpaca_connect('paper',
                         Sys.getenv("ALPACA_PAPER_KEY"),
                         Sys.getenv("ALPACA_PAPER_SECRET"))
d_conn <- alpaca_connect('data',
                         Sys.getenv("ALPACA_LIVE_KEY"),
                         Sys.getenv("ALPACA_LIVE_SECRET"))

portfolio_model <- read_portfolio_model(system.file(package='rblncr','extdata/sample_portfolio.yaml'))

get_portfolio_current(t_conn) |>
  load_portfolio_targets(portfolio_model) |>
  price_portfolio(connection = d_conn, price_type = 'close') |>
  solve_portfolio() |>
  constrain_orders(d_conn) |>
  trader(trading_connection = t_conn,
         pricing_connection = d_conn,
         verbose = FALSE)
#>              timestamp symbol order  limit filled   status
#> 1  2024-03-11 16:20:35   AAPL   -58 172.93    -58   filled
#> 2  2024-03-11 16:20:36   GOOG    73 139.00      0 canceled
#> 3  2024-03-11 16:20:36     VT    91 108.56      0 canceled
#> 4                 <NA>   TSLA    28     NA      0 no_limit
#> 5  2024-03-11 16:20:46   GOOG    73 138.99      0 canceled
#> 6  2024-03-11 16:20:46     VT    91 108.57     91   filled
#> 7                 <NA>   TSLA    28     NA      0 no_limit
#> 8  2024-03-11 16:20:55   GOOG    73 139.77     73   filled
#> 9                 <NA>   TSLA    28     NA      0 no_limit
#> 10                <NA>   TSLA    28     NA      0 no_limit
#> 11                <NA>   TSLA    28     NA      0 no_limit
#> 12                <NA>   TSLA    28     NA      0 no_limit
#> 13                <NA>   TSLA    28     NA      0 no_limit
#> 14                <NA>   TSLA    28     NA      0 no_limit
#> 15                <NA>   TSLA    28     NA      0 no_limit
#> 16                <NA>   TSLA    28     NA      0 no_limit
#> 17                <NA>   TSLA    28     NA      0 no_limit
#> 18                <NA>   TSLA    28     NA      0 no_limit
#> 19                <NA>   TSLA    28     NA      0 no_limit
#> 20                <NA>   TSLA    28     NA      0 no_limit
#> 21                <NA>   TSLA    28     NA      0 no_limit
#> 22                <NA>   TSLA    28     NA      0 no_limit
#> 23                <NA>   TSLA    28     NA      0 no_limit
#> 24                <NA>   TSLA    28     NA      0 no_limit
#> 25                <NA>   TSLA    28     NA      0 no_limit
#> 26 2024-03-11 16:23:27   TSLA    28 178.78     28   filled